Extract property edit parts
This extracts the property edit parts from the PropertyTable class and
moves them into a separate package.
